Fix time issues in season
I was getting an error that the global `time` was not defined. That was
fixed by importing time.

Then, I got an error where there wasn't a `strftime` attribute on time.
datetime objects have `strftime`, but time doesn't. I tried switching
this to use datetime instead of time at first, but that led to some
weird issues when calling strptime multiple times. So I went with using
time consistently instead.